+++ src/ypkgupgr/consts.py
@@ -1,0 +1,2 @@
+APP_NAME = "ypkgupgr"
+AUTHOR_NAME = "Yesser Studios"
