+++ src/ypkgupgr/misc.py
@@ -1,0 +1,10 @@
+failed = ""
+outdated_count = 0
+finished_count = 0
+
+ypkgupgr_outdated = False
+
+ran_from_script = False
+
+line_length = dict()
+line_count = 0
